+++ contracts/interfaces/IPoolManagerLogic.sol
@@ -8,61 +8,61 @@ interface IPoolManagerLogic {
         bool useForDeposits;
     }
 
-    /**
-    * @dev Given the address of an asset, returns whether this pool can hold the asset.
+     /**
+    * @notice Returns whether this pool can hold the given asset.
     * @param _asset Address of the asset.
     * @return bool Whether this pool can hold the asset.
     */
     function isAvailableAsset(address _asset) external view returns (bool);
 
     /**
-    * @dev Given the address of an asset, returns whether the pool accepts the asset for deposits.
+    * @notice Returns whether the pool accepts the given asset for deposits.
     * @param _asset Address of the asset.
     * @return bool Whether this pool can accept the asset for deposits.
     */
     function isDepositAsset(address _asset) external view returns (bool);
 
     /**
-    * @dev Returns a list of assets that can be deposited into the pool.
+    * @notice Returns a list of assets that can be deposited into the pool.
     */
     function getDepositAssets() external view returns (address[] memory);
 
     /**
-    * @dev Returns a list of assets that the pool can hold.
+    * @notice Returns a list of assets that the pool can hold.
     */
     function getAvailableAssets() external view returns (address[] memory);
 
     /**
-    * @dev Returns the pool's performance fee.
+    * @notice Returns the pool's performance fee.
     */
-    function performanceFee() external view returns (uint);
+    function performanceFee() external view returns (uint256);
 
     /**
-    * @dev Adds a new asset to the list of acceptable assets for deposits.
+    * @notice Adds a new asset to the list of acceptable assets for deposits.
     * @param _asset Address of the asset.
     */
     function addDepositAsset(address _asset) external;
 
     /**
-    * @dev Removes an asset from the list of acceptable assets for deposits.
+    * @notice Removes an asset from the list of acceptable assets for deposits.
     * @param _asset Address of the asset.
     */
     function removeDepositAsset(address _asset) external;
 
     /**
-    * @dev Updates the pool's performance fee.
+    * @notice Updates the pool's performance fee.
     * @param _performanceFee The new performance fee.
     */
-    function setPerformanceFee(uint _performanceFee) external;
+    function setPerformanceFee(uint256 _performanceFee) external;
 
     /**
-    * @dev Adds a new asset to the list of assets the pool can hold.
+    * @notice Adds a new asset to the list of assets the pool can hold.
     * @param _asset Address of the asset.
     */
     function addAvailableAsset(address _asset) external;
 
     /**
-    * @dev Removes an asset from the list of assets the pool can hold.
+    * @notice Removes an asset from the list of assets the pool can hold.
     * @param _asset Address of the asset.
     */
     function removeAvailableAsset(address _asset) external;
